Refactor #3 - move Advanced Search view to its own file
- We also moved the shared `has_filters` method to utils. (Tidying up utils.py is probably next on our list when we're done with views)
